Register a NameService for better TLS support
A limitation of DCR at present is that to Java, your SSL-equipped
service exists at the hostname 192.168.99.101 or whatever, and not the
docker-compose hostname (`db` or whatever).

This frequently leads to irritating hacks in order for normal clients to
not error - replacing the SSLSocketFactory and the HostnameValidator is
enough with OkHttp - but you frequently don't have this choice - for
example the Multipass and Gatekeeper clients force you onto their TLS
configurations. In any case, you don't want these properties to be set
in your integration tests.

Proposed workarounds usually involve either not using TLS (which means
it can't be an ETE test), using Dash etc - which has considerable
configurational cost, or alternatively getting a cert for `localhost`,
`db`, and whatever IPs your docker machine might end up on. This is not
ideal.

Instead, this commit maintains a `NameService`, a class that Java can be
configured with to use instead of its internal DNS resolver - think of
it as a Java specific `/etc/hosts`.

It's use is entirely optional, to use it the client merely needs to set
the system property `sun.net.spi.nameservice.provider.1` to
`dns,docker-compose-rule`. When the hostname requested is not a
container, it delegates to the system dns infrastructure.

If this system property is set, the tester can communicate with the
server using the container name as the hostname.