Ignore test that fails on Windows/Mac by design.
This commit uses JUnit's assumptions to ignore the
HostNetworkedPortsIntegrationTest on any OS but linux.

Ports exposed on the "host" network are accessible on linux because the host
is running on localhost. On Windows and Mac, the host is instead running on a
linux VM and the ports are not accessible.